+++ tests/conftest.py
@@ -2,29 +2,35 @@ import pathlib
 
 import pytest
 
-import pyflp
+from pyflp import Project, parse
+from pyflp.mixer import Mixer
 
 
 @pytest.fixture(scope="session")
 def project():
-    return pyflp.parse(pathlib.Path(__file__).parent / "assets" / "FL 20.8.4.flp")
+    return parse(pathlib.Path(__file__).parent / "assets" / "FL 20.8.4.flp")
 
 
 @pytest.fixture(scope="session")
-def arrangements(project: pyflp.Project):
+def arrangements(project: Project):
     return project.arrangements
 
 
 @pytest.fixture(scope="session")
-def rack(project: pyflp.Project):
+def rack(project: Project):
     return project.channels
 
 
 @pytest.fixture(scope="session")
-def mixer(project: pyflp.Project):
+def mixer(project: Project):
     return project.mixer
 
 
 @pytest.fixture(scope="session")
-def patterns(project: pyflp.Project):
+def inserts(mixer: Mixer):
+    return tuple(mixer)[:25]
+
+
+@pytest.fixture(scope="session")
+def patterns(project: Project):
     return project.patterns